pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "speechmux-plugin-stt"
version = "0.1.0"
description = "SpeechMux STT Inference Plugin base framework (servicer, Protocol, dummy engine)"
requires-python = ">=3.10"
dependencies = [
    "grpcio>=1.60.0",
    "protobuf>=4.25.0",
    "pyyaml>=6.0",
    # proto generated package — for local dev: pip install -e ../proto/gen/python
    # "speechmux-proto @ git+ssh://git@github.com/speechmux/proto.git@v0.1.0#subdirectory=gen/python",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "ruff>=0.4",
    "mypy>=1.10",
    "grpcio-tools>=1.60.0",
    "types-PyYAML>=6.0",
]

[project.entry-points."speechmux.stt_engine"]
dummy = "speechmux_plugin_stt.engine.dummy:DummySTTEngine"

[tool.hatch.build.targets.wheel]
packages = ["src/speechmux_plugin_stt"]

[tool.ruff]
line-length = 100

[tool.mypy]
python_version = "3.10"
strict = true
